+++ src/FreeElephants/JsonApiToolkit/Routing/FastRoute/EmptyOperationIdPatchedDispatcherFactory.php
@@ -4,8 +4,12 @@ namespace FreeElephants\JsonApiToolkit\Routing\FastRoute;
 
 use cebe\openapi\spec\Operation;
 use cebe\openapi\spec\PathItem;
+use FastRoute\DataGenerator;
 use FastRoute\Dispatcher;
 use FastRoute\RouteCollector;
+use FastRoute\RouteParser;
+use FreeElephants\JsonApiToolkit\JsonApi\Request\Route\RouteFactory;
+use FreeElephants\JsonApiToolkit\Routing\FastRoute\Dispatcher\JsonApiDispatcher;
 use function FastRoute\simpleDispatcher;
 use FreeElephants\JsonApiToolkit\OasToolsAdapter\OpenApiDocumentParserInterface;
 use FreeElephants\JsonApiToolkit\OasToolsAdapter\YamlStringParser;
@@ -17,30 +21,49 @@ class EmptyOperationIdPatchedDispatcherFactory implements DispatcherFactoryInter
 {
     private OpenApiDocumentParserInterface $apiDocumentParser;
     private OperationHandlerNormalizerInterface $operationHandlerNormalizer;
+    private RouteFactory $routeFactory;
 
-    public function __construct(OperationHandlerNormalizerInterface $operationHandlerNormalizer = null, OpenApiDocumentParserInterface $apiDocumentParser = null)
+    public function __construct(
+        RouteFactory $routeFactory,
+        OperationHandlerNormalizerInterface $operationHandlerNormalizer = null,
+        OpenApiDocumentParserInterface $apiDocumentParser = null
+    )
     {
         $this->operationHandlerNormalizer = $operationHandlerNormalizer ?: new DefaultOperationHandlerNormalizer();
         $this->apiDocumentParser = $apiDocumentParser ?: new YamlStringParser();
+        $this->routeFactory = $routeFactory;
     }
 
     public function buildDispatcher(string $openApiDocumentSource): Dispatcher
     {
         $openapi = $this->apiDocumentParser->parse($openApiDocumentSource);
         $paths = $openapi->paths->getIterator();
-        $dispatcher = simpleDispatcher(function (RouteCollector $routeCollector) use ($paths) {
-            /** @var PathItem $pathItem */
-            foreach ($paths as $path => $pathItem) {
-                /** @var Operation $operation */
-                foreach ($pathItem->getOperations() as $method => $operation) {
-                    $httpMethod = $this->normalizeHttpMethod($method);
-                    $handler = $this->normalizeOperationHandler($operation);
-                    $routeCollector->addRoute($httpMethod, $path, $handler);
-                }
-            }
-        });
 
-        return $dispatcher;
+        $routeCollector = $this->buildRouteCollector();
+        foreach ($paths as $path => $pathItem) {
+            foreach ($pathItem->getOperations() as $method => $operation) {
+                $httpMethod = $this->normalizeHttpMethod($method);
+                $handler = $this->normalizeOperationHandler($operation);
+                $routeCollector->addRoute($httpMethod, $path, $handler);
+            }
+        }
+
+        return new JsonApiDispatcher($routeCollector->getData(), $this->routeFactory);
+    }
+
+    private function buildDataGenerator(): DataGenerator
+    {
+        return new DataGenerator\GroupCountBased();
+    }
+
+    private function buildRouteParser(): RouteParser
+    {
+        return new RouteParser\Std();
+    }
+
+    private function buildRouteCollector(): RouteCollector
+    {
+        return new RouteCollector($this->buildRouteParser(), $this->buildDataGenerator());
     }
 
     private function normalizeHttpMethod(string $method): string